+++ CHANGELOG.md
@@ -6,10 +6,10 @@
 * Added `triggerImmediately` to `onChange` method.
 * Report real-time progress information about downloads through `SyncStatus.downloadProgress`.
 * Compose: Add `composeState()` extension method on `SyncStatus`.
-* Add `includeOld` option on `Table` which sets `CrudEntry.oldData` to previous values on updates.
-* Add `includeMetadata` option on `Table` which adds a `_metadata` column that can be used for updates.
+* Add `trackPreviousValues` option on `Table` which sets `CrudEntry.oldData` to previous values on updates.
+* Add `trackMetadata` option on `Table` which adds a `_metadata` column that can be used for updates.
   The configured metadata is available through `CrudEntry.metadata`.
-* Add `ignoreEmptyUpdate` option which skips creating CRUD entries for updates that don't change any values.
+* Add `ignoreEmptyUpdates` option which skips creating CRUD entries for updates that don't change any values.
 
 ## 1.0.0-BETA32